+++ src/base/b-virtual-scroll/modules/chunk-render.ts
@@ -221,7 +221,7 @@ export default class ChunkRender extends Friend {
 		const
 			{ctx, chunkRequest} = this;
 
-		if (ctx.dataProvider == null && ctx.options.length === 0) {
+		if (ctx.dataProvider == null && ctx.items.length === 0) {
 			return;
 		}