+++ CHANGELOG.md
@@ -13,7 +13,11 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.0.0-rc.??? (2021-01-??)
 
-#### :house: Internal
+#### :boom: Breaking Change
+
+* Now all `option(*)` fields for the trait `iItems` are renamed to `item(*)` and marked as deprecated for
+  * `base/b-virtual-scroll`
+  * `base/b-slider`
 
 * `base/b-matryoshka` marked as deprecated. Use `base/b-tree` instead.